fix query-replace-defaults compatibility with emacs 24
In Emacs 24, this was only a (cons from to). In Emacs 25, it became a
list of (cons from to).

This patch separates the defaults from vr and Emacs' builtin replace, so
search/replacing in one does not affect the defaults display of the
other.

In Emacs 25, the two remain fully compatible.